feat(consensus): fast sync test with pending subgag (#9979)
# Description of change

Implement a test scenario where the validator has some pending subdags
before the fast sync starts to work, i.e. last solid commit index is
smaller than the last commit index. It causes panic in the current
implementation. The fix is coming soon.


Also moves fast sync test with voting blocks to fast.rs

## Links to any relevant issues

Fixes issue #9978.

## How the change has been tested

- [x] Basic tests (linting, compilation, formatting, unit/integration
tests)
- [x] Patch-specific tests (correctness, functionality coverage)
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [x] I have checked that new and existing unit tests pass locally with
my changes
