+++ crates/starfish/core/src/commit_syncer/fast.rs
@@ -101,6 +101,10 @@ impl<C: NetworkClient> FastCommitSyncer<C> {
             sync_type: CommitSyncType::Fast,
         });
         let last_solid_commit_index = inner.dag_state.read().last_solid_commit_index();
+        info!(
+            "[fast_commit_sync] Initialized with synced_commit_index={}",
+            last_solid_commit_index
+        );
         FastCommitSyncer {
             inner,
             inflight_fetches: JoinSet::new(),
@@ -781,5 +785,747 @@ impl<C: NetworkClient> FastCommitSyncer<C> {
     #[allow(dead_code)]
     fn synced_commit_index(&self) -> CommitIndex {
         self.synced_commit_index
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use std::time::{Duration, Instant};
+
+    use iota_protocol_config::ProtocolConfig;
+    use prometheus::Registry;
+    use starfish_config::{Parameters, local_committee_and_keys};
+    use tempfile::TempDir;
+    use tokio::time::sleep;
+    use tracing::info;
+    use typed_store::DBMetrics;
+
+    use crate::authority_node::tests::make_authority_with_params;
+
+    /// Test that voting blocks stored during fast sync can be served to peers.
+    /// This test verifies:
+    /// 1. Validator A fast syncs and stores voting block headers
+    /// 2. Validator B fast syncs and can receive commits/voting blocks from A
+    /// 3. Both validators agree on commit history
+    ///
+    /// Test flow to ensure B requests commits A has in voting storage:
+    /// - Phase 1: All run → commits 1-N1 (all validators have these)
+    /// - Phase 2: Stop B first (B stops at N1)
+    /// - Phase 3: A + the other 5 validators continue → commits N1-N2 (B
+    ///   doesn't have these)
+    /// - Phase 4: Stop A (A stops at N2)
+    /// - Phase 5: The remaining 5 validators continue → commits N2-N3 (neither
+    ///   A nor B have these)
+    /// - Phase 6: Restart A, fast syncs N2-N3 → stores voting headers
+    /// - Phase 7: Restart B, needs N1-N3 → should get N2-N3 from A's voting
+    ///   storage
+    #[tokio::test(flavor = "current_thread")]
+    async fn test_fast_sync_voting_blocks_served_to_peer() {
+        telemetry_subscribers::init_for_testing();
+        let db_registry = Registry::new();
+        DBMetrics::init(&db_registry);
+
+        // Use 7 validators so that quorum (5) can still be reached with 2 validators
+        // stopped.
+        const NUM_AUTHORITIES: usize = 7;
+        const COMMIT_GAP_THRESHOLD: u32 = 30;
+
+        // Work phases need to be long enough to create a gap larger than
+        // COMMIT_GAP_THRESHOLD (30) for fast sync to trigger.
+        let stable_work_duration = Duration::from_secs(10);
+
+        let (committee, keypairs) = local_committee_and_keys(0, vec![1; NUM_AUTHORITIES]);
+        let mut protocol_config = ProtocolConfig::get_for_max_version_UNSAFE();
+        protocol_config.set_consensus_transaction_ref_for_testing(true);
+
+        let temp_dirs: Vec<TempDir> = (0..NUM_AUTHORITIES)
+            .map(|_| TempDir::new().unwrap())
+            .collect();
+
+        let mut authorities = Vec::with_capacity(NUM_AUTHORITIES);
+        let mut boot_counters = [0u64; NUM_AUTHORITIES];
+        let mut consumer_monitors = Vec::with_capacity(NUM_AUTHORITIES);
+        let mut output_receivers = Vec::with_capacity(NUM_AUTHORITIES);
+
+        let validator_a_index: usize = 0;
+        let validator_b_index: usize = 1;
+
+        // Start all authorities
+        for (index, _) in committee.authorities() {
+            let parameters = Parameters {
+                db_path: temp_dirs[index.value()].path().to_path_buf(),
+                dag_state_cached_rounds: 5,
+                commit_sync_parallel_fetches: 2,
+                commit_sync_batch_size: 10,
+                commit_sync_gap_threshold: COMMIT_GAP_THRESHOLD,
+                fast_commit_sync_batch_size: 20,
+                sync_last_known_own_block_timeout: Duration::from_millis(2_000),
+                ..Default::default()
+            };
+            let (authority, receiver, monitor) = make_authority_with_params(
+                index,
+                &temp_dirs[index.value()],
+                committee.clone(),
+                keypairs.clone(),
+                boot_counters[index],
+                protocol_config.clone(),
+                parameters,
+                0,
+            )
+            .await;
+            boot_counters[index] += 1;
+            authorities.push(authority);
+            output_receivers.push(receiver);
+            consumer_monitors.push(monitor);
+        }
+
+        // Phase 1: Let all authorities run and commit transactions
+        let start_time = Instant::now();
+        let mut committed_index = [0u32; NUM_AUTHORITIES];
+        while start_time.elapsed() < stable_work_duration {
+            for (index, receiver) in output_receivers.iter_mut().enumerate() {
+                while let Ok(committed_subdag) = receiver.try_recv() {
+                    let commit_index = committed_subdag.commit_ref.index;
+                    assert!(
+                        commit_index > committed_index[index],
+                        "Commit index {} should be greater than previous {}",
+                        commit_index,
+                        committed_index[index]
+                    );
+                    committed_index[index] = commit_index;
+                    consumer_monitors[index].set_highest_handled_commit(commit_index);
+                }
+            }
+            sleep(Duration::from_millis(50)).await;
+        }
+
+        // Phase 2: Stop validator B first (so B misses commits created while it's down)
+        let last_processed_b = consumer_monitors[validator_b_index].highest_handled_commit();
+        authorities.remove(validator_b_index).stop().await;
+
+        // Phase 3: Let A and others continue committing (B misses these)
+        let start_time = Instant::now();
+        while start_time.elapsed() < stable_work_duration {
+            for (index, receiver) in output_receivers.iter_mut().enumerate() {
+                if index == validator_b_index {
+                    continue;
+                }
+                while let Ok(committed_subdag) = receiver.try_recv() {
+                    let commit_index = committed_subdag.commit_ref.index;
+                    assert!(
+                        commit_index > committed_index[index],
+                        "Commit index {} should be greater than previous {}",
+                        commit_index,
+                        committed_index[index]
+                    );
+                    committed_index[index] = commit_index;
+                    consumer_monitors[index].set_highest_handled_commit(commit_index);
+                }
+            }
+            sleep(Duration::from_millis(50)).await;
+        }
+
+        // Phase 4: Stop validator A
+        let last_processed_a = consumer_monitors[validator_a_index].highest_handled_commit();
+        authorities.remove(validator_a_index).stop().await;
+
+        // Phase 5: Let the remaining validators (all except A and B) continue
+        // committing (both A and B miss these commits)
+        let start_time = Instant::now();
+        while start_time.elapsed() < stable_work_duration {
+            for (index, receiver) in output_receivers.iter_mut().enumerate() {
+                if index == validator_a_index || index == validator_b_index {
+                    continue;
+                }
+                while let Ok(committed_subdag) = receiver.try_recv() {
+                    let commit_index = committed_subdag.commit_ref.index;
+                    assert!(
+                        commit_index > committed_index[index],
+                        "Commit index {} should be greater than previous {}",
+                        commit_index,
+                        committed_index[index]
+                    );
+                    committed_index[index] = commit_index;
+                    consumer_monitors[index].set_highest_handled_commit(commit_index);
+                }
+            }
+            sleep(Duration::from_millis(50)).await;
+        }
+
+        // Phase 6: Restart validator A - it will fast sync and store voting blocks
+        let parameters = Parameters {
+            db_path: temp_dirs[validator_a_index].path().to_path_buf(),
+            dag_state_cached_rounds: 5,
+            commit_sync_parallel_fetches: 2,
+            commit_sync_batch_size: 10,
+            commit_sync_gap_threshold: COMMIT_GAP_THRESHOLD,
+            fast_commit_sync_batch_size: 20,
+            sync_last_known_own_block_timeout: Duration::from_millis(2_000),
+            ..Default::default()
+        };
+        let (authority, receiver, monitor) = make_authority_with_params(
+            committee.to_authority_index(validator_a_index).unwrap(),
+            &temp_dirs[validator_a_index],
+            committee.clone(),
+            keypairs.clone(),
+            boot_counters[validator_a_index],
+            protocol_config.clone(),
+            parameters,
+            last_processed_a,
+        )
+        .await;
+        boot_counters[validator_a_index] += 1;
+        output_receivers[validator_a_index] = receiver;
+        consumer_monitors[validator_a_index] = monitor;
+        authorities.insert(validator_a_index, authority);
+
+        // Wait for validator A to catch up via fast sync
+        let start_time = Instant::now();
+        let mut a_caught_up = false;
+        while start_time.elapsed() < Duration::from_secs(60) {
+            for (index, receiver) in output_receivers.iter_mut().enumerate() {
+                if index == validator_b_index {
+                    continue;
+                }
+                while let Ok(committed_subdag) = receiver.try_recv() {
+                    let commit_index = committed_subdag.commit_ref.index;
+                    assert!(
+                        commit_index > committed_index[index],
+                        "Commit index {} should be greater than previous {}",
+                        commit_index,
+                        committed_index[index]
+                    );
+                    committed_index[index] = commit_index;
+                    consumer_monitors[index].set_highest_handled_commit(commit_index);
+                }
+            }
+
+            let a_index = consumer_monitors[validator_a_index].highest_handled_commit();
+            let max_other = consumer_monitors
+                .iter()
+                .enumerate()
+                .filter(|(i, _)| *i != validator_a_index && *i != validator_b_index)
+                .map(|(_, m)| m.highest_handled_commit())
+                .max()
+                .unwrap_or(0);
+
+            if a_index > 0 && a_index + 20 >= max_other {
+                a_caught_up = true;
+                break;
+            }
+            sleep(Duration::from_millis(50)).await;
+        }
+
+        assert!(
+            a_caught_up,
+            "Validator A should have caught up via fast sync"
+        );
+
+        // Phase 7: Restart validator B - it needs commits that A fast-synced
+        // B should be able to get voting block headers from A's voting storage
+        let parameters = Parameters {
+            db_path: temp_dirs[validator_b_index].path().to_path_buf(),
+            dag_state_cached_rounds: 5,
+            commit_sync_parallel_fetches: 2,
+            commit_sync_batch_size: 10,
+            commit_sync_gap_threshold: COMMIT_GAP_THRESHOLD,
+            fast_commit_sync_batch_size: 20,
+            sync_last_known_own_block_timeout: Duration::from_millis(2_000),
+            ..Default::default()
+        };
+        let (authority, receiver, monitor) = make_authority_with_params(
+            committee.to_authority_index(validator_b_index).unwrap(),
+            &temp_dirs[validator_b_index],
+            committee.clone(),
+            keypairs.clone(),
+            boot_counters[validator_b_index],
+            protocol_config.clone(),
+            parameters,
+            last_processed_b,
+        )
+        .await;
+        output_receivers[validator_b_index] = receiver;
+        consumer_monitors[validator_b_index] = monitor;
+
+        authorities.insert(validator_b_index, authority);
+
+        // Wait for validator B to catch up
+        let start_time = Instant::now();
+        let mut b_caught_up = false;
+        while start_time.elapsed() < Duration::from_secs(60) {
+            for (index, receiver) in output_receivers.iter_mut().enumerate() {
+                while let Ok(committed_subdag) = receiver.try_recv() {
+                    let commit_index = committed_subdag.commit_ref.index;
+                    assert!(
+                        commit_index > committed_index[index],
+                        "Commit index {} should be greater than previous {}",
+                        commit_index,
+                        committed_index[index]
+                    );
+                    committed_index[index] = commit_index;
+                    consumer_monitors[index].set_highest_handled_commit(commit_index);
+                }
+            }
+
+            let b_index = consumer_monitors[validator_b_index].highest_handled_commit();
+            let max_other = consumer_monitors
+                .iter()
+                .enumerate()
+                .filter(|(i, _)| *i != validator_b_index)
+                .map(|(_, m)| m.highest_handled_commit())
+                .max()
+                .unwrap_or(0);
+
+            if b_index > 0 && b_index + 20 >= max_other {
+                b_caught_up = true;
+                break;
+            }
+            sleep(Duration::from_millis(50)).await;
+        }
+
+        assert!(
+            b_caught_up,
+            "Validator B should have caught up via fast sync"
+        );
+
+        // Verify both validators A and B have similar commit indices
+        let a_final = consumer_monitors[validator_a_index].highest_handled_commit();
+        let b_final = consumer_monitors[validator_b_index].highest_handled_commit();
+
+        assert!(
+            a_final > last_processed_a,
+            "Validator A should have progressed: before_restart={}, final={}",
+            last_processed_a,
+            a_final
+        );
+        assert!(
+            b_final > last_processed_b,
+            "Validator B should have progressed: before_restart={}, final={}",
+            last_processed_b,
+            b_final
+        );
+
+        // Both should be within reasonable range of each other
+        let diff = (a_final as i64 - b_final as i64).unsigned_abs() as u32;
+        assert!(
+            diff < 30,
+            "Validators A and B should have similar commit indices: A={}, B={}, diff={}",
+            a_final,
+            b_final,
+            diff
+        );
+
+        // Collect voting block headers metrics to verify voting storage was used.
+        let total_hits: u64 = authorities
+            .iter()
+            .map(|a| {
+                a.context()
+                    .metrics
+                    .node_metrics
+                    .commit_sync_voting_block_headers_hits
+                    .get()
+            })
+            .sum();
+
+        let total_fallbacks: u64 = authorities
+            .iter()
+            .map(|a| {
+                a.context()
+                    .metrics
+                    .node_metrics
+                    .commit_sync_voting_block_headers_fallbacks
+                    .get()
+            })
+            .sum();
+
+        info!(
+            "Voting block headers metrics: hits={}, fallbacks={}",
+            total_hits, total_fallbacks
+        );
+
+        // In tests, peer selection is deterministic (not shuffled), so B will
+        // request from A first. A has voting storage for commits it fast-synced,
+        // so we should get voting hits.
+        assert!(
+            total_hits > 0,
+            "Expected voting block headers hits > 0, got {}",
+            total_hits
+        );
+
+        let commit_sync_fetch_commits_handler_uncertified_skipped: u64 = authorities
+            .iter()
+            .map(|a| {
+                a.context()
+                    .metrics
+                    .node_metrics
+                    .commit_sync_fetch_commits_handler_uncertified_skipped
+                    .with_label_values(&["fast_commit_sync"])
+                    .get()
+            })
+            .sum();
+
+        assert!(
+            commit_sync_fetch_commits_handler_uncertified_skipped > 0,
+            "Expected uncertified commits skipped > 0 for fast sync, got {}",
+            commit_sync_fetch_commits_handler_uncertified_skipped
+        );
+
+        // Stop all authorities
+        for authority in authorities {
+            authority.stop().await;
+        }
+    }
+
+    /// Test that a validator with pending subdags (gap between last_commit and
+    /// last_solid_commit_leader_round) can successfully catch up via fast sync
+    /// after restart.
+    ///
+    /// This test creates pending subdags using dynamic peer unsubscribe, then
+    /// stops and restarts the validator to verify fast sync handles
+    /// pre-existing pending subdags correctly.
+    ///
+    /// Test flow:
+    /// - Phase 1: All validators run together, creating initial commits
+    /// - Phase 2: Dynamically unsubscribe test validator from validator 1 +
+    ///   stop txn synchronizer + stop shard reconstructor
+    /// - Phase 3: Wait for commits with missing txs (creates pending subdags)
+    ///   and verify gap
+    /// - Phase 4: Stop test validator
+    /// - Phase 5: Other validators continue (creates fast sync gap > threshold)
+    /// - Phase 6: Restart test validator with full connectivity, but keep txn
+    ///   synchronizer + shard reconstructor stopped to prevent pending subdags
+    ///   from being solidified
+    /// - Phase 7: Verify fast sync was used and validator caught up
+    #[tokio::test(flavor = "current_thread")]
+    async fn test_fast_sync_with_pending_subdags() {
+        telemetry_subscribers::init_for_testing();
+        let db_registry = Registry::new();
+        DBMetrics::init(&db_registry);
+
+        const NUM_AUTHORITIES: usize = 4;
+        const COMMIT_GAP_THRESHOLD: u32 = 30;
+        const COMMIT_SYNC_BATCH_SIZE: u32 = 20;
+
+        // Work phases need to be long enough to create pending subdags during Phase 3.
+        // During Phase 3, the validator keeps creating commits (headers arrive via
+        // cordial dissemination), so there's no commit gap for syncers to act
+        // on. Phase 5 creates a commit gap larger than the threshold for fast
+        // sync to trigger on restart.
+        let stable_work_duration = Duration::from_secs(10);
+
+        let (committee, keypairs) = local_committee_and_keys(0, vec![1; NUM_AUTHORITIES]);
+        let mut protocol_config = ProtocolConfig::get_for_max_version_UNSAFE();
+        protocol_config.set_consensus_transaction_ref_for_testing(true);
+
+        let temp_dirs: Vec<TempDir> = (0..NUM_AUTHORITIES)
+            .map(|_| TempDir::new().unwrap())
+            .collect();
+
+        let mut authorities = Vec::with_capacity(NUM_AUTHORITIES);
+        let mut boot_counters = [0u64; NUM_AUTHORITIES];
+        let mut consumer_monitors = Vec::with_capacity(NUM_AUTHORITIES);
+        let mut output_receivers = Vec::with_capacity(NUM_AUTHORITIES);
+
+        let test_validator_index: usize = 0;
+        let blocked_validator_index: usize = 1;
+
+        // Phase 1: Start all authorities and let them create initial commits
+        for (index, _) in committee.authorities() {
+            let parameters = Parameters {
+                db_path: temp_dirs[index.value()].path().to_path_buf(),
+                dag_state_cached_rounds: 5,
+                commit_sync_parallel_fetches: 2,
+                commit_sync_batch_size: COMMIT_SYNC_BATCH_SIZE,
+                commit_sync_gap_threshold: COMMIT_GAP_THRESHOLD,
+                fast_commit_sync_batch_size: COMMIT_SYNC_BATCH_SIZE,
+                sync_last_known_own_block_timeout: Duration::from_millis(2_000),
+                ..Default::default()
+            };
+            let (authority, receiver, monitor) = make_authority_with_params(
+                index,
+                &temp_dirs[index.value()],
+                committee.clone(),
+                keypairs.clone(),
+                boot_counters[index],
+                protocol_config.clone(),
+                parameters,
+                0,
+            )
+            .await;
+            boot_counters[index] += 1;
+            authorities.push(authority);
+            output_receivers.push(receiver);
+            consumer_monitors.push(monitor);
+        }
+
+        let mut txn_counter = 0u64;
+        let start_time = Instant::now();
+        let mut committed_index = [0u32; NUM_AUTHORITIES];
+        while start_time.elapsed() < stable_work_duration {
+            // Submit transactions to all validators (rotating)
+            let authority_index = txn_counter as usize % authorities.len();
+            let txn = vec![txn_counter as u8; 16];
+            authorities[authority_index]
+                .transaction_client()
+                .submit(vec![txn])
+                .await
+                .unwrap();
+            txn_counter += 1;
+
+            for (index, receiver) in output_receivers.iter_mut().enumerate() {
+                while let Ok(committed_subdag) = receiver.try_recv() {
+                    let commit_index = committed_subdag.commit_ref.index;
+                    assert!(
+                        commit_index > committed_index[index],
+                        "Commit index {} should be greater than previous {}",
+                        commit_index,
+                        committed_index[index]
+                    );
+                    committed_index[index] = commit_index;
+                    consumer_monitors[index].set_highest_handled_commit(commit_index);
+                }
+            }
+            sleep(Duration::from_millis(50)).await;
+        }
+
+        // Phase 2: Dynamically unsubscribe from validator 1 + stop txn synchronizer +
+        // stop shard reconstructor This will create pending subdags as headers
+        // arrive via cordial dissemination but transactions from validator 1's
+        // blocks are missing and shards cannot be reconstructed. Commit syncers
+        // won't activate during Phase 3 because there's no commit gap - the
+        // validator keeps up with commits, just missing transactions.
+        authorities[test_validator_index].unsubscribe_from_peer_for_test(
+            committee
+                .to_authority_index(blocked_validator_index)
+                .unwrap(),
+        );
+        authorities[test_validator_index]
+            .stop_transactions_synchronizer_for_test()
+            .await
+            .expect("Transaction synchronizer should stop");
+        authorities[test_validator_index]
+            .stop_shard_reconstructor_for_test()
+            .await
+            .expect("Shard reconstructor should stop");
+
+        // Phase 3: Wait for headers to arrive via cordial dissemination and commits to
+        // be created. Submit transactions to all validators (rotating).
+        // This should create pending subdags (gap between
+        // last_commit and last_solid_commit_leader_round)
+
+        // Track commits before the wait
+        let commits_before = committed_index[test_validator_index];
+
+        // Submit transactions to all validators during Phase 3.
+        // Validator 0 can process transactions from itself and validators 2 & 3.
+        // However, validator 0 can't fetch transactions from validator 1's blocks
+        // because:
+        // - It's unsubscribed from validator 1
+        // - Transaction synchronizer is stopped (blocks active transaction fetching)
+        // - Shard reconstructor is stopped (blocks erasure-coded shard reconstruction)
+        // This creates pending subdags.
+        // Commit syncers are running but have nothing to fetch (no commit gap exists).
+        let phase3_start = Instant::now();
+        while phase3_start.elapsed() < stable_work_duration {
+            // Submit transactions to all validators (rotating)
+            let authority_index = txn_counter as usize % authorities.len();
+            let txn = vec![txn_counter as u8; 16];
+            authorities[authority_index]
+                .transaction_client()
+                .submit(vec![txn])
+                .await
+                .unwrap();
+            txn_counter += 1;
+
+            // Drain receivers
+            for (index, receiver) in output_receivers.iter_mut().enumerate() {
+                while let Ok(committed_subdag) = receiver.try_recv() {
+                    let commit_index = committed_subdag.commit_ref.index;
+                    if commit_index > committed_index[index] {
+                        committed_index[index] = commit_index;
+                        consumer_monitors[index].set_highest_handled_commit(commit_index);
+                    }
+                }
+            }
+            sleep(Duration::from_millis(50)).await;
+        }
+
+        let commits_after = committed_index[test_validator_index];
+        let new_commits = commits_after - commits_before;
+        assert!(
+            new_commits > 0,
+            "Expected new commits during Phase 3, got 0"
+        );
+
+        // Verify pending subdags gap exists
+        let dag_state = authorities[test_validator_index].dag_state_for_test();
+        let last_commit = dag_state.read().last_commit_round();
+        let last_solid = dag_state.read().last_solid_commit_leader_round();
+        let dag_round = dag_state.read().threshold_clock_round();
+
+        // Verify gap exists - we expect pending subdags
+        let has_gap = last_commit > last_solid.unwrap_or(0);
+        assert!(
+            has_gap,
+            "Expected pending subdags gap: last_commit={}, last_solid={:?}. \
+             DAG round={}, new_commits={}. \
+             Validator 1's new blocks should have missing transactions.",
+            last_commit, last_solid, dag_round, new_commits
+        );
+
+        // Record where the validator is now (with pending subdags)
+        let last_processed_with_pending =
+            consumer_monitors[test_validator_index].highest_handled_commit();
+
+        // Phase 4: Stop test validator (preserves pending subdags to disk)
+        authorities.remove(test_validator_index).stop().await;
+
+        // Phase 5: Let other validators continue while the test validator is stopped
+        // (creates fast sync gap > threshold)
+        let start_time = Instant::now();
+        while start_time.elapsed() < stable_work_duration * 2 {
+            for (index, receiver) in output_receivers.iter_mut().enumerate() {
+                if index == test_validator_index {
+                    continue; // Skip stopped validator
+                }
+                while let Ok(committed_subdag) = receiver.try_recv() {
+                    let commit_index = committed_subdag.commit_ref.index;
+                    if commit_index > committed_index[index] {
+                        committed_index[index] = commit_index;
+                        consumer_monitors[index].set_highest_handled_commit(commit_index);
+                    }
+                }
+            }
+            sleep(Duration::from_millis(50)).await;
+        }
+
+        let max_other = consumer_monitors
+            .iter()
+            .enumerate()
+            .filter(|(i, _)| *i != test_validator_index)
+            .map(|(_, m)| m.highest_handled_commit())
+            .max()
+            .unwrap_or(0);
+
+        let gap = max_other.saturating_sub(last_processed_with_pending);
+        assert!(
+            gap > COMMIT_GAP_THRESHOLD,
+            "Gap {} should be greater than threshold {}",
+            gap,
+            COMMIT_GAP_THRESHOLD
+        );
+
+        // Phase 6: Restart test validator with full connectivity.
+        // Transaction synchronizer and shard reconstructor
+        // will be stopped after restart to prevent pending subdags from being
+        // solidified.
+        let parameters = Parameters {
+            db_path: temp_dirs[test_validator_index].path().to_path_buf(),
+            dag_state_cached_rounds: 5,
+            commit_sync_parallel_fetches: 2,
+            commit_sync_batch_size: COMMIT_SYNC_BATCH_SIZE,
+            commit_sync_gap_threshold: COMMIT_GAP_THRESHOLD,
+            fast_commit_sync_batch_size: COMMIT_SYNC_BATCH_SIZE,
+            sync_last_known_own_block_timeout: Duration::from_millis(2_000),
+            ..Default::default()
+        };
+        let (authority, receiver, monitor) = make_authority_with_params(
+            committee.to_authority_index(test_validator_index).unwrap(),
+            &temp_dirs[test_validator_index],
+            committee.clone(),
+            keypairs.clone(),
+            boot_counters[test_validator_index],
+            protocol_config.clone(),
+            parameters,
+            last_processed_with_pending,
+        )
+        .await;
+        output_receivers[test_validator_index] = receiver;
+        consumer_monitors[test_validator_index] = monitor;
+        authorities.insert(test_validator_index, authority);
+
+        // Keep transaction synchronizer and shard reconstructor stopped after restart
+        // to prevent pending subdags from being solidified. This forces the system
+        // to rely on fast sync to fill the gap, which should expose the bug.
+        authorities[test_validator_index]
+            .stop_transactions_synchronizer_for_test()
+            .await
+            .expect("Transaction synchronizer should stop");
+        authorities[test_validator_index]
+            .stop_shard_reconstructor_for_test()
+            .await
+            .expect("Shard reconstructor should stop");
+
+        // Phase 7: Wait for the validator to catch up via fast sync
+        // This tests whether fast sync can handle pre-existing pending subdags
+        let start_time = Instant::now();
+        let mut caught_up = false;
+        while start_time.elapsed() < Duration::from_secs(60) {
+            for (index, receiver) in output_receivers.iter_mut().enumerate() {
+                while let Ok(committed_subdag) = receiver.try_recv() {
+                    let commit_index = committed_subdag.commit_ref.index;
+                    if commit_index > committed_index[index] {
+                        committed_index[index] = commit_index;
+                        consumer_monitors[index].set_highest_handled_commit(commit_index);
+                    }
+                }
+            }
+
+            let test_index = consumer_monitors[test_validator_index].highest_handled_commit();
+            let max_other = consumer_monitors
+                .iter()
+                .enumerate()
+                .filter(|(i, _)| *i != test_validator_index)
+                .map(|(_, m)| m.highest_handled_commit())
+                .max()
+                .unwrap_or(0);
+
+            if test_index > last_processed_with_pending && test_index + 20 >= max_other {
+                caught_up = true;
+                break;
+            }
+            sleep(Duration::from_millis(50)).await;
+        }
+
+        assert!(
+            caught_up,
+            "Validator {} should have caught up via fast sync",
+            test_validator_index
+        );
+
+        // Verify the validator progressed significantly after restart with pending
+        // subdags
+        let final_index = consumer_monitors[test_validator_index].highest_handled_commit();
+        assert!(
+            final_index > last_processed_with_pending,
+            "Validator should have progressed after restart: with_pending_subdags={}, final={}",
+            last_processed_with_pending,
+            final_index
+        );
+
+        // Verify that fast sync was actually used by checking the fetched commits
+        // metric
+        let commit_sync_fetched_commits: u64 = authorities
+            .iter()
+            .map(|a| {
+                a.context()
+                    .metrics
+                    .node_metrics
+                    .commit_sync_fetched_commits
+                    .with_label_values(&["fast_commit_sync"])
+                    .get()
+            })
+            .sum();
+
+        assert!(
+            commit_sync_fetched_commits > 0,
+            "Expected commits fetched via fast sync > 0, got {}",
+            commit_sync_fetched_commits
+        );
+
+        // Stop all authorities
+        for authority in authorities {
+            authority.stop().await;
+        }
     }
 }
